+++ Model.CRUD.PCL/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+﻿using System.Resources;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("Model CRUD Portable")]
+[assembly: AssemblyDescription("Platform Model CRUD Pattern (Portable)")]
+
+[assembly: AssemblyCulture("")]