+++ Model.CRUD/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("Model CRUD")]
+[assembly: AssemblyDescription("Platform Model CRUD Pattern")]
+
+[assembly: Guid("a57d24e5-511a-4db2-944e-fc9d7aa01b67")]