+++ Model.CRUD/Types/Operations.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Platform.Model.CRUD.Types
+{
+    public enum Operations
+    {
+        Unknow = -1,
+        None = 0,
+        Create = 1,
+        Read = 2,
+        Update = 3, 
+        Delete = 4
+    }
+}