+++ Model.SPS.PlugIn/Properties/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+﻿using System.Resources;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("Model SPS PlugIn Portable")]
+[assembly: AssemblyDescription("Platform Model SPS PlugIn Pattern (Portable)")]
+
+[assembly: AssemblyCulture("")]