+++ Model.SPS/Interfaces/IPlugIn.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Platform.Model.SPS.Interfaces
+{
+    /// <summary>
+    /// This interface is implemented all plugins.
+    /// PlugIns does not directly implement this interface, but they implement by inheriting PlugIn class.
+    /// </summary>
+    public interface IPlugIn : IPluggable
+    {
+        /// <summary>
+        /// Name of the plugin.
+        /// </summary>
+        string Name { get; }
+    }
+
+}